% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comm.R
\name{comm}
\alias{comm}
\title{Microbiome Subnetwork Classification and Central Bacterial Identification}
\usage{
comm(
  community,
  phylum,
  algorithm = "walktrap",
  nsub = 9,
  SubM = NA,
  organism = NA
)
}
\arguments{
\item{community}{A microbial relative abundance matrix. Each row represents a microbe(species or genus) and each column represents a sample.}

\item{phylum}{Corresponding relationship of different taxa.}

\item{algorithm}{Community partitioning algorithm, default is "walktrap". Other optional parameters are "clusters", "edge.betweenness", "label.propagation", "leading.eigenvector", "fastgreedy" or "multilevel"}

\item{nsub}{The number of candidate subnetworks. Default is 9.}

\item{SubM}{Key subnetwork ID selected according to the evaluation results of "netcore.R". Default is NA.}

\item{organism}{The name of the microbe selected according to the evaluation results of "nodescore.R". Default is NA.}
}
\value{
This function returns a list with the following components:
\item{corr_net.pdf}{Correlation network visualization results.}
\item{corrnet_character.csv}{Correlation network topological properties table.}
\item{class_subnetwork.csv}{The result of community partition.}
\item{size_tax_subnetwork.csv}{The size of the subnetworks.}
\item{netclass_allsubnet.pdf}{Community partition visualization results.}
\item{subnet_character.csv}{Candidate subnetworks topological properties table}
\item{selected_subnet.pdf}{Selected subnetwork visualization results}
\item{nodes_character_selectedsubnet.csv}{Topological properties of the nodes in the selected subnetwork.}
\item{centralnode_net.pdf}{Central node network visualization results.}
}
\description{
Microbiome Subnetwork Classification and Central Bacterial Identification
}
\references{
Yihua Wang (2023): NetClass: A noncontrol-dependent approach for microbiome subnetwork classification and central bacterial identification.
}
