% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodesocre.R
\name{NodeScore}
\alias{NodeScore}
\title{Evaluating and scoring central node microbes}
\usage{
NodeScore(df1)
}
\arguments{
\item{df1}{the topological attributes of the chosen subnetwork nodes (microbes) using five network centrality measures}
}
\value{
This function returns a list with the following components:
\item{nodes_score.csv}{the topological attributes scores and rank of the chosen subnetwork nodes (microbes) using five network centrality measures.}
}
\description{
Evaluating and scoring central node microbes
}
